matrix-app-front/src/core/modules/auth/enrollment/domain/use-cases/verify-current-user-attribute/verify-current-user-attribute.use-case.test.ts
import { IEnrollmentRepository } from '../../../repository/enrollment.repository';
import VerifyCurrentUserAttributeUseCase from './verify-current-user-attribute.use-case';

const mockRepository = {
  verifyUserAttribute: jest.fn(),
} as unknown as IEnrollmentRepository;

const verifyCurrentUserAttributeUseCase = new VerifyCurrentUserAttributeUseCase(mockRepository);

describe('VerifyCurrentUserAttributeUseCase', () => {
  describe('execute', () => {
    it('should call repository.sendVerifyCodeAttribute with the correct data', async () => {
      const inputData = { attribute: 'email', code: 'ASB2VD' };
      const response = await verifyCurrentUserAttributeUseCase.execute(inputData);
      expect(response).toBeUndefined();
      expect(mockRepository.verifyUserAttribute).toHaveBeenCalledWith(inputData);
    });
  });
});
